backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp>=2024.1.1
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
slowapi==0.1.9

